requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-text-splitters
qdrant-client
requests
tiktoken
PyMuPDF
python-dotenv